% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{best_alignment}
\alias{best_alignment}
\title{Find the best alignment}
\usage{
best_alignment(scoringMat, tracebackMat)
}
\arguments{
\item{scoringMat}{filled Scoring Matrix obtained from \code{\link{fill_matrices}} applying the Smith-Waterman algorithm.}

\item{tracebackMat}{filled Traceback Matrix obtained from \code{\link{fill_matrices}} applying the Smith-Waterman algorithm.}
}
\value{
A list of the two aligned sequences, the score of the optimal alignment and a matrix with starting ending position.
}
\description{
Starting from the element with the highest score, this function performs the traceback
based on the source of each score recursively, until 0 is encountered.
The function returns the optimal local alignment between the two sequences and its score.
}
