% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fill_matrices}
\alias{fill_matrices}
\title{Fill the Scoring Matrix and Traceback Matrix}
\usage{
fill_matrices(scoringMat, tracebackMat, seqA, seqB, match, mismatch, gap)
}
\arguments{
\item{scoringMat}{Initialized empty Scoring Matrix of integers, with the first row and first column set to 0.}

\item{tracebackMat}{Initialized empty Traceback Matrix of characters, with the first row and column set to 0.}

\item{seqA}{A vector of the first nucleotide sequence with "-" as first element.}

\item{seqB}{A vector of the second nucleotide sequence with "-" as first element.}

\item{match}{A positive number associated with a nucleotide match.}

\item{mismatch}{A negative number associated with a nucleotide mismatch.}

\item{gap}{A negative number associated with the gap penalty.}
}
\value{
A list with the filled Scoring Matrix and the filled Traceback Matrix.
}
\description{
For each pair of nucleotides, the function assigns a score using the Smith Waterman greedy algorithm.
It inserts the value in the scoring matrix and, depending on the score,
it stores the direction of the alignment in the traceback matrix.
}
