% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pairwise_score}
\alias{pairwise_score}
\title{Calculate Pairwise Score}
\usage{
pairwise_score(nt1, nt2, match, mismatch)
}
\arguments{
\item{nt1}{Nucleotide of the first sequence, a character.}

\item{nt2}{Nucletide of the second sequence, a character.}

\item{match}{A positive number associated with a nucleotide match.}

\item{mismatch}{A negative number associated with a nucleotide mismatch.}
}
\value{
An integer representing the score associated to the comparison of the two nucleotides.
}
\description{
This function assigns to a pair of nucleotides a score, whether the two nucleotides match or not.
The value depends on the parameters match and mismatch, which are part of the scoring system decided a priori.
}
